chore: temp rename Avatar.tsx
